+++ config/application.rb
@@ -41,5 +41,12 @@ module AtuservicioRails
       components/site_header_cell
       components/story_slider_cell
     )
+
+    api_mime_types = %W(
+      application/vnd.api+json
+      text/x-json
+      application/json
+    )
+    Mime::Type.register 'application/vnd.api+json', :json, api_mime_types
   end
 end